fix the types of several params in RDLS,VOF,VOF3P, and VOS3P

+++ proteus/mprans/RDLS.py
@@ -1059,7 +1059,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
         argsDict["useTimeIntegration"] = useTimeIntegration
         argsDict["lag_shockCapturing"] = self.shockCapturing.lag
         argsDict["lag_subgridError"] = self.stabilization.lag
-        argsDict["shockCapturingDiffusion"] = self.shockCapturing.shockCapturingFactor
+        argsDict["shockCapturingDiffusion"] = float(self.shockCapturing.shockCapturingFactor)
         argsDict["u_l2g"] = self.u[0].femSpace.dofMap.l2g
         argsDict["elementDiameter"] = self.elementDiameter
         argsDict["nodeDiametersArray"] = self.mesh.nodeDiametersArray
@@ -1169,7 +1169,7 @@ class LevelModel(proteus.Transport.OneLevelTransport):
         argsDict["useTimeIntegration"] = useTimeIntegration
         argsDict["lag_shockCapturing"] = self.shockCapturing.lag
         argsDict["lag_subgridError"] = int(self.stabilization.lag)
-        argsDict["shockCapturingDiffusion"] = int(self.shockCapturing.shockCapturingFactor)
+        argsDict["shockCapturingDiffusion"] = float(self.shockCapturing.shockCapturingFactor)
         argsDict["u_l2g"] = self.u[0].femSpace.dofMap.l2g
         argsDict["elementDiameter"] = self.elementDiameter
         argsDict["nodeDiametersArray"] = self.mesh.nodeDiametersArray
